Refactor : lotDetail / bid atom 분리
