+++ gokkan/src/store/lotDetailAtom.ts
@@ -1,5 +1,4 @@
 import { atomsWithQuery } from 'jotai-tanstack-query';
-import axios from 'axios';
 import { atom } from 'jotai';
 import { Client } from '@stomp/stompjs';
 import { getCustomAxios } from '../utils/customAxios';
@@ -36,17 +35,6 @@ interface ImageUrl {
   url: string;
 }
 
-interface AuctionInfo {
-  auctionEndDateTime: string;
-  currentPrice: number;
-}
-
-interface BidInfo {
-  memberId: string;
-  price: number;
-  bidTime: string;
-}
-
 interface ExpertValuation {
   name: string;
   profileImageUrl: string;
@@ -62,118 +50,48 @@ interface SellerInfo {
   createdAt: string;
 }
 
-const baseUrl = 'http://3.38.59.40:8080/api/v1';
-
 const lotIdAtom = atom(7);
-const auctionIdAtom = atom(2);
 
 const [lotDetailAtom] = atomsWithQuery((get) => ({
   queryKey: ['lotDetail', get(lotIdAtom)],
   queryFn: async ({ queryKey: [, lotId] }): Promise<LotDetail> => {
     const customAxios = getCustomAxios();
-    const res = await customAxios({
-      method: 'get',
-      url: `${baseUrl}/items/details/auction`,
+    const { data } = await customAxios.get('/api/v1/items/details/auction', {
       params: {
         itemId: lotId,
       },
     });
 
-    return res?.data;
-  },
-}));
-
-const [auctionInfoAtom] = atomsWithQuery((get) => ({
-  queryKey: ['bidInfo', get(auctionIdAtom)],
-  queryFn: async ({ queryKey: [, auctionId] }): Promise<AuctionInfo> => {
-    const res = await axios({
-      method: 'get',
-      url: `${baseUrl}/auction`,
-      params: {
-        auctionId: auctionId,
-      },
-    });
-
-    return res?.data;
-  },
-}));
-
-const [bidHistoryAtom] = atomsWithQuery((get) => ({
-  queryKey: ['bidHistory', get(auctionIdAtom)],
-  queryFn: async ({ queryKey: [, auctionId] }): Promise<BidInfo[] | null[]> => {
-    const res = await axios({
-      method: 'get',
-      url: `${baseUrl}/auction/history`,
-      params: {
-        auctionId: auctionId,
-      },
-    });
-
-    return res?.data;
+    return data;
   },
 }));
 
 const [expertValuationAtom] = atomsWithQuery((get) => ({
   queryKey: ['expertValuation', get(lotIdAtom)],
   queryFn: async ({ queryKey: [, lotId] }): Promise<ExpertValuation> => {
-    const res = await axios({
-      method: 'get',
-      url: `${baseUrl}/expert/comment`,
+    const customAxios = getCustomAxios();
+    const { data } = await customAxios.get('/api/v1/expert/comment', {
       params: {
         itemId: lotId,
       },
     });
 
-    return res?.data;
+    return data;
   },
 }));
 
 const [sellerInfoAtom] = atomsWithQuery((get) => ({
   queryKey: ['sellerInfo', get(lotIdAtom)],
   queryFn: async ({ queryKey: [, lotId] }): Promise<SellerInfo> => {
-    const res = await axios({
-      method: 'get',
-      url: `${baseUrl}/users/seller`,
+    const customAxios = getCustomAxios();
+    const { data } = await customAxios.get('/api/v1/users/seller', {
       params: {
         itemId: lotId,
       },
     });
 
-    return res?.data;
+    return data;
   },
 }));
 
-const StompClientAtom = atom<React.MutableRefObject<Client | undefined> | null>(
-  null
-);
-
-// const bidCloseTimeAtom = atom((get) => {
-//   const { auctionEndDateTime } = get(auctionInfoAtom);
-//   return auctionEndDateTime;
-// });
-
-const currBidPriceAtom = atom<number | string>(0);
-
-const currBidHistoryAtom = atom<BidInfo[] | null[]>([]);
-
-const bidCloseTimeAtom = atom('');
-
-const bidErrMsgAtom = atom('');
-
-const addedBidTimeAtom = atom('');
-
-export {
-  lotIdAtom,
-  auctionIdAtom,
-  lotDetailAtom,
-  auctionInfoAtom,
-  bidHistoryAtom,
-  expertValuationAtom,
-  sellerInfoAtom,
-  StompClientAtom,
-  currBidPriceAtom,
-  bidCloseTimeAtom,
-  bidErrMsgAtom,
-  addedBidTimeAtom,
-  currBidHistoryAtom,
-};
+export { lotIdAtom, lotDetailAtom, expertValuationAtom, sellerInfoAtom };
